+++ src/proto_alpha/lib_protocol/test/integration/consensus/test_participation.ml
@@ -69,9 +69,16 @@ let bake_and_attest_once (_b_pred, b_cur) baker attester =
 let test_participation ~sufficient_participation () =
   let open Lwt_result_wrap_syntax in
   let n_accounts = 2 in
-  let* b0, accounts =
-    Context.init_n ~consensus_threshold_size:1 n_accounts ()
+  let constants =
+    let c = Default_parameters.constants_test in
+    {
+      c with
+      dal = {c.dal with rewards_ratio = Q.(0 // 100)};
+      issuance_weights = {c.issuance_weights with dal_rewards_weight = 0};
+      consensus_threshold_size = 1;
+    }
   in
+  let* b0, accounts = Context.init_with_constants_n constants n_accounts in
   let* csts = Context.get_constants (B b0) in
   let blocks_per_cycle = Int32.to_int csts.parametric.blocks_per_cycle in
   let mpr = csts.parametric.minimal_participation_ratio in
@@ -174,9 +181,7 @@ let check_no_dal_participation
     Assert.equal_bool ~loc:__LOC__ dal_info.sufficient_dal_participation true
   in
   let* () =
-    (* No Tez are actually provisioned for DAL rewards as incentives are not
-       enabled in this test. Turn the test to [Assert.not_equal_tez] when
-       incentives are enabled by default. *)
+    (* No Tez are provisioned for DAL rewards when [dal_rewards_weight = 0]. *)
     Assert.equal_tez ~loc:__LOC__ dal_info.expected_dal_rewards Tez.zero
   in
   return_unit
@@ -187,9 +192,15 @@ let check_no_dal_participation
 let test_participation_rpc () =
   let open Lwt_result_wrap_syntax in
   let n_accounts = 2 in
-  let* b0, (account1, account2) =
-    Context.init2 ~consensus_threshold_size:1 ()
+  let constants =
+    let c = Default_parameters.constants_test in
+    {
+      c with
+      issuance_weights = {c.issuance_weights with dal_rewards_weight = 0};
+      consensus_threshold_size = 1;
+    }
   in
+  let* b0, (account1, account2) = Context.init_with_constants2 constants in
   let del1 = Context.Contract.pkh account1 in
   let del2 = Context.Contract.pkh account2 in
   let* csts = Context.get_constants (B b0) in